move to using nirtorch
